Crate distance

A collection of approximate string matching algorithms

This library contains algorithms dealing with approximate string matching. These algorithms can be used to tell the approximate difference between two strings. This is usful for a varity of things like spell checking, fuzzy search, etc.

Algorithms

Levenshtein distance

Damerau–Levenshtein distance

Hamming distance

Sift3 distance

Enums

DistanceError
distance error type

Functions

damerau_levenshtein
Calculates the Damerau-Levenshtein distance between two strings.

hamming
Calculates the Hamming distance between two strings of equal length.

levenshtein
Calculates the Levenshtein distance between two strings.

sift3
Calculates the sift3 distance between two strings with a default max_distance of 5.